Feat: Backup (#2945)
New backups

This PR contains the code of the new backups functionality. Written in Kotlin, called from Scala.
RFC: https://docs.google.com/document/d/14XUnGmPpvDIwiSI4NB6LAcG30P07i7fvrL6kgA1j2aQ/edit#heading=h.8o1gv9hkmgwl
Encryption rewritten in Sodium v1, according to: https://wearezeta.atlassian.net/wiki/spaces/PROD/pages/59965445/Exporting+history+v1

+++ storage/src/main/kotlin/com/waz/zclient/storage/db/conversations/ConversationsEntity.kt
@@ -35,28 +35,28 @@ data class ConversationsEntity(
     val managed: Boolean?,
 
     @ColumnInfo(name = "last_event_time", defaultValue = "0")
-    val lastEventTime: Int,
+    val lastEventTime: Long,
 
     @ColumnInfo(name = "is_active", defaultValue = "0")
     val active: Boolean,
 
     @ColumnInfo(name = "last_read", defaultValue = "0")
-    val lastRead: Int,
+    val lastRead: Long,
 
     @ColumnInfo(name = "muted_status", defaultValue = "0")
     val mutedStatus: Int,
 
     @ColumnInfo(name = "mute_time", defaultValue = "0")
-    val muteTime: Int,
+    val muteTime: Long,
 
     @ColumnInfo(name = "archived", defaultValue = "0")
     val archived: Boolean,
 
     @ColumnInfo(name = "archive_time", defaultValue = "0")
-    val archiveTime: Int,
+    val archiveTime: Long,
 
     @ColumnInfo(name = "cleared")
-    val cleared: Int?,
+    val cleared: Long?,
 
     @ColumnInfo(name = "generated_name", defaultValue = "")
     val generatedName: String,
@@ -83,10 +83,10 @@ data class ConversationsEntity(
     val verified: String?,
 
     @ColumnInfo(name = "ephemeral")
-    val ephemeral: Int?,
+    val ephemeral: Long?,
 
     @ColumnInfo(name = "global_ephemeral")
-    val globalEphemeral: Int?,
+    val globalEphemeral: Long?,
 
     @ColumnInfo(name = "unread_call_count", defaultValue = "0")
     val unreadCallCount: Int,
